% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_nearest_neighbor.R
\name{generate_four_nearest_neighbor_matrix}
\alias{generate_four_nearest_neighbor_matrix}
\title{Generate adjacency matrix for a four-nearest-neighbor regular lattice}
\usage{
generate_four_nearest_neighbor_matrix(v, k)
}
\arguments{
\item{v}{Number of locations in the rectangular field; Must be an even number}

\item{k}{Number of rows in the rectangular field with a total of \code{v} locations.}
}
\value{
A v*v adjacency matrix
}
\description{
This function generates the adjacency matrix for a regular rectangular lattice.
Each location on the lattice has four neighbors on orthogonal directions
The neighbors wraps around on edges of the lattice.
}
\details{
for example if \code{v = 30},\code{k = 5}, then the indices of the locations are as follows.

1  6 11 16 21 26

2  7 12 17 22 27

3  8 13 18 23 28

4  9 14 19 24 29

5 10 15 20 25 30
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
