% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{get_eta_param_space_small}
\alias{get_eta_param_space_small}
\title{Compute the parameter space of eta from the adjacency matrix}
\usage{
get_eta_param_space_small(adj_mat)
}
\arguments{
\item{adj_mat}{A (transformed) symmetric adjacency matrix.}
}
\value{
A vector of the lower and upper bounds for eta
}
\description{
The bounds of eta is decided by the reciprocal of the maximum and minimum of eigenvalues of the adjacency matrix.
}
\details{
This is for small matrices that R can handle with its memory.
If your matrix is big (n >10000), consider the \code{preprocess_big} function.
The input matrix should be transformed using \code{transform_small()} or \code{preprocess_big}
to be used in \code{pmrf()} directly as the \code{bounds_e} argument.
}
\examples{
mat = matrix(c(1,1,0,1,1,0,0,0,1), nrow = 3)
get_eta_param_space(mat)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
