% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{get_jump_frequency}
\alias{get_jump_frequency}
\title{Percentage of jumps in MCMC iterations}
\usage{
get_jump_frequency(ret, Ti, N)
}
\arguments{
\item{ret}{returned list from \code{pmrf()}}

\item{Ti}{number of total iterations}

\item{N}{datasize}
}
\value{
A list with jump percentages for \code{w}, \code{alpha}, \code{eta} and \code{tau2}
}
\description{
Percentage of jumps in MCMC iterations
}
\examples{
adj_trans = transform_small(irregular_lattice)
n = dim(irregular_lattice)[1]
y = simulate_y(n = n, adj_mat = adj_trans, alpha = 2, eta = 2, tau2 = 3, m = 2, M = 2000)
bounds_e = get_eta_param_space_small(adj_trans)
ret = pmrf(total_iter = 1000, N = n , y, adj_trans, c(1, 0.5, NA, 2), bounds_e)
get_jump_frequency(ret, 1000, n)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
