% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irregular_lattice}
\alias{irregular_lattice}
\title{Irregular lattice}
\format{A matrix with 1000 rows and 1000 columns}
\usage{
irregular_lattice
}
\description{
An adjacency matrix for 1000 locations (nodes) with 1997 edges among them.
}
\keyword{datasets}
