% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmrf.R
\name{pmrf}
\alias{pmrf}
\title{Poisson mixed Markov Random Field model inference}
\usage{
pmrf(total_iter, N, y, adj_mat, vars, bounds_e, bounds_a = c(-10, 10),
  bounds_t = c(0, 10), inis = c(0.1, 0, 0.1), wInis = rnorm(N,
  inis[1], sqrt(inis[3])))
}
\arguments{
\item{total_iter}{Total number of MCMC iterations.}

\item{N}{Number of locations that data is observed. Multiple observations can be made at each location.}

\item{y}{Observed data at each location. \code{y} should be a matrix with \code{N} rows, and number of columns equal to the number of observations at each location.}

\item{adj_mat}{Transformed adjacency matrix for the locations. See Details on how to transform the adjacency matrix.
Must be in the 'dgCMatrix' class. Use \code{as(yourmatrix, 'dgCMatrix')} to convert.}

\item{vars}{Variances for the random walk proposal for the MCMC for each parameter,
\code{w}, \code{alpha}, \code{eta} and \code{tau2}.}

\item{bounds_e}{Lower and upper bound for \code{eta}. This should be decided by the eigenvalues of the adjacency matrix. See details. No default supplied.}

\item{bounds_a}{Lower and upper bound for \code{alpha}. Default \code{c(-10,10)}.}

\item{bounds_t}{Lower and upper bound for \code{tau} (Not \code{tau2}, see Details for the prior distributions.)
Default \code{c(0,10)}.}

\item{inis}{Initial values for \code{alpha}, \code{eta} and \code{tau2}. Default \code{c(0.1,0.0,0.1)}.}

\item{wInis}{Initial values for \code{w}. Must be a vector of length \code{N}. Default \code{rnorm(N, inis[1], sqrt(inis[3]))}.}
}
\value{
A list with the posterior results.  \code{w} is a \code{N} by \code{total_iters} matrix,
\code{alpha}, \code{eta} and \code{tau2} are vectors with length \code{N} respectively.
Last element in the list is a vector of length four that counts the number of jumps made for each parameter.
}
\description{
A function to perform a double Metropolis-Hastings algorithm for
the parameter estimates of spatial count data observed on a network under a Poisson hierarchical Markov Random Field model.
The parameters are \code{w}, \code{alpha}, \code{eta} and \code{tau2}. \code{alpha} is connected with the conditional mean,
\code{eta} is connected with the level of spatial dependency and \code{tau2} is connected with conditional variance.
See References for detailed model specification.
}
\details{
The adjacency matrix \code{adj_mat} needs to be transformed by dividing each nonzero element (default is \code{1}) by the sum of the row sum and column sum.
See  and  for  ways to transform.

The parameter space of \code{eta} is defined as the reciprocal of the maximum and minimum of the eigenvalues of the (transformed) adjacency matrix.
(i.e. \code{c(1/min(eigen(adj_mat)$values), 1/max(eigen(adj_mat)$values))})
See and for ways to compute.

The \code{vars} argument supplies the variance used in the random walk proposal.
In general, the larger the variance, the smaller the jump frequency.

For \code{eta}, an independence proposal was used. So the third element of \code{vars} is never used.

For \code{alpha} and \code{eta}, uniform priors were with the specified bounds were used for each.
For \code{tau2}, a uniform prior is specified for \code{tau},
and the prior distribution for \code{tau2} is derived using change-of-variables.

You can supply the inital values using the last iteration of a previous MCMC run to continue to process.
}
\examples{
pmrf(5000, nrow(y), y, adj_mat, c(0.5,0.5,NA,1), c(-2,2))
pmrf(2000, nrow(y), y, adj_mat, c(0.5,0.5,NA,1), c(-2,2), inis = c(0,0,0), wInis = rnorm(nrow(y), 0, 1))

}
\references{
Zhou, N., Friedberg, I. & Kaiser, M.S. (2019).  ""

Liang, F. (2010). A double Metropolis–Hastings sampler for spatial models with intractable normalizing constants. \emph{Journal of Statistical Computation and Simulation}, 80(9), 1007-1022.
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
