% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{preprocess_big}
\alias{preprocess_big}
\title{Matrix transformation and eigenvalue calculation for large matrices}
\usage{
preprocess_big(adj_mat, savepath)
}
\arguments{
\item{adj_mat}{An \strong{untransformed} symmetric adjacency matrix.}

\item{savepath}{The path to save the transformed matrix and bounds_eta.
If \code{NULL} or missing, then not saving.}
}
\value{
A list of \code{b_eta}: parameter space of eta and \code{new_mat}: transformed matrix.
}
\description{
This function has the combined function of
 \code{transform_small()} and \code{get_eta_param_space_small}
}
\details{
This function is designed for large matrices that R cannot handle.
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
