% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{print_param_estimates}
\alias{print_param_estimates}
\title{Print parameter estimates to console}
\usage{
print_param_estimates(ret, burn_in = 200, digit = 3)
}
\arguments{
\item{ret}{returned list from \code{pmrf()}}

\item{burn_in}{number of iterations to throw away. Default 200.}

\item{digit}{number of digits after the decimal point to display. Default 3.}
}
\description{
Parameter estimates include mean and 95\% credible interval.
}
\details{
It is STRONGLY recommended that you output the estimates after throwing away burn-in iterations.
}
\examples{
adj_trans = transform_small(irregular_lattice)
n = dim(irregular_lattice)[1]
y = simulate_y(n = n, adj_mat = adj_trans, alpha = 2, eta = 2, tau2 = 3, m = 2, M = 2000)
bounds_e = get_eta_param_space_small(adj_trans)
ret = pmrf(total_iter = 1000, N = n , y, adj_trans, c(1, 0.5, NA, 2), bounds_e)
print_param_estimates(ret, 300, 4)
print_param_estimates(ret, 0, 2)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
