% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regular_lattice}
\alias{regular_lattice}
\title{Square regular lattice}
\format{A matrix with 900 rows and 900 columns}
\usage{
regular_lattice
}
\description{
An adjacency matrix for a 30*30 square lattice with 900 locations
}
\keyword{datasets}
