% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_y}
\alias{simulate_y}
\title{Simulate data with given parameters}
\usage{
simulate_y(n, adj_mat, alpha, eta, tau2, m, M)
}
\arguments{
\item{n}{Number of locations.}

\item{adj_mat}{\code{n*n} adjacency matrix.}

\item{alpha}{True parameter value for \code{alpha}.}

\item{eta}{True parameter value for \code{eta}}

\item{tau2}{True parameter value for \code{tau2}}

\item{m}{Number of observations per location.}

\item{M}{Number of Gibbs iterations}
}
\value{
A \code{n*m} matrix of data y
}
\description{
Simulate data with given parameters
}
\examples{
adj_trans = transform_small(irregular_lattice)
y = simulate_y(dim(irregular_lattice)[1], adj_trans, alpha = 2, eta = 2, tau2 = 3, m = 2, M = 2000)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
