% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{transform_small}
\alias{transform_small}
\title{Transform adjacency matrix by dividing sum of row sum and column sum}
\usage{
transform_small(adj_mat)
}
\arguments{
\item{adj_mat}{A symmetric adjacency matrix.
Edges are represented as 1 and non edges 0.}
}
\value{
A transformed matrix in the \code{dgCMatrix} class
}
\description{
Transform adjacency matrix by dividing sum of row sum and column sum
}
\details{
This is for small matrices that R can handle with its memory.
If your matrix is big (n >10000), consider the \code{preprocess_big} function.
}
\examples{
mat = matrix(c(1,1,0,1,1,0,0,0,1), nrow = 3)
transform_small(mat)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
