% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{delete_burn_in}
\alias{delete_burn_in}
\title{Delete burn in iterations in MCMC}
\usage{
delete_burn_in(ret, burn_in)
}
\arguments{
\item{ret}{returned list from MCMC}

\item{burn_in}{number of iterations to throw away}

\item{N}{datasize}
}
\value{
A list with MCMC iterations after burn in for \code{w}, \code{alpha}, \code{eta} and \code{tau2}
}
\description{
Delete burn in iterations in MCMC
}
\examples{
delete_burn_in(ret, 400)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
