% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{get_jump_frequency}
\alias{get_jump_frequency}
\title{Percentage of jumps in MCMC iterations}
\usage{
get_jump_frequency(ret, Ti, N)
}
\arguments{
\item{ret}{returned list from \code{pmrf()}}

\item{Ti}{number of total iterations}

\item{N}{datasize}
}
\value{
A list with jump percentages for \code{w}, \code{alpha}, \code{eta} and \code{tau2}
}
\description{
Percentage of jumps in MCMC iterations
}
\examples{
get_jump_frequency(ret, 10000, 200)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
