% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{print_param_estimates}
\alias{print_param_estimates}
\title{Print parameter estimates to console}
\usage{
print_param_estimates(ret, burn_in = 200, digit = 3)
}
\arguments{
\item{ret}{returned list from \code{pmrf()}}

\item{burn_in}{number of iterations to throw away. Default 200.}

\item{digit}{number of digits after the decimal point to display. Default 3.}
}
\description{
Parameter estimates include mean and 95\% credible interval.
}
\details{
It is STRONGLY recommended that you output the estimates after throwing away burn-in iterations.
}
\examples{
print_param_estimates(ret, 500, 4)
print_param_estimates(ret, 0, 2)
}
\author{
Naihui Zhou (ashley.n.zhou@gmail.com)
}
